refactor: Separate class
