+++ feature/explore/src/main/java/com/teamoffroad/feature/explore/presentation/ExploreCameraViewModel.kt
@@ -3,7 +3,7 @@ package com.teamoffroad.feature.explore.presentation
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.teamoffroad.feature.explore.domain.usecase.PostExploreQrAuthUseCase
-import com.teamoffroad.feature.explore.presentation.model.ExploreCameraUiState
+import com.teamoffroad.feature.explore.presentation.model.ExploreResultState
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
@@ -16,15 +16,15 @@ class ExploreCameraViewModel @Inject constructor(
     private val postExploreQrAuthUseCase: PostExploreQrAuthUseCase,
 ) : ViewModel() {
 
-    private val _uiState: MutableStateFlow<ExploreCameraUiState> = MutableStateFlow(ExploreCameraUiState.None)
-    val uiState: StateFlow<ExploreCameraUiState> = _uiState.asStateFlow()
+    private val _exploreResultState: MutableStateFlow<ExploreResultState> = MutableStateFlow(ExploreResultState.None)
+    val exploreResultState: StateFlow<ExploreResultState> = _exploreResultState.asStateFlow()
 
     private val _resultImageUrl: MutableStateFlow<String> = MutableStateFlow("")
     val resultImageUrl: StateFlow<String> = _resultImageUrl.asStateFlow()
 
     fun postExploreResult(placeId: Long, latitude: Double, longitude: Double, qr: String) {
-        if (uiState.value is ExploreCameraUiState.Loading) return
-        _uiState.value = ExploreCameraUiState.Loading
+        if (exploreResultState.value is ExploreResultState.Loading) return
+        _exploreResultState.value = ExploreResultState.Loading
 
         viewModelScope.launch {
             runCatching {
@@ -33,18 +33,18 @@ class ExploreCameraViewModel @Inject constructor(
                 when (exploreResult.isQRMatched) {
                     true -> {
                         _resultImageUrl.value = exploreResult.characterImageUrl
-                        _uiState.value = ExploreCameraUiState.Success
+                        _exploreResultState.value = ExploreResultState.Success
                     }
 
                     false -> {
                         _resultImageUrl.value = exploreResult.characterImageUrl
-                        _uiState.value = ExploreCameraUiState.CodeError
+                        _exploreResultState.value = ExploreResultState.CodeError
                     }
                 }
             }.onFailure {
                 when (it.message) {
-                    "HTTP 400" -> _uiState.value = ExploreCameraUiState.LocationError
-                    else -> _uiState.value = ExploreCameraUiState.EtcError
+                    "HTTP 400" -> _exploreResultState.value = ExploreResultState.LocationError
+                    else -> _exploreResultState.value = ExploreResultState.EtcError
                 }
             }
         }
