+++ feature/explore/src/main/java/com/teamoffroad/feature/explore/presentation/component/ExploreOffroadMap.kt
@@ -1,0 +1,259 @@
+package com.teamoffroad.feature.explore.presentation.component
+
+import android.view.Gravity
+import androidx.compose.foundation.background
+import androidx.compose.foundation.gestures.detectTapGestures
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.navigationBarsPadding
+import androidx.compose.foundation.layout.offset
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.input.pointer.pointerInput
+import androidx.compose.ui.layout.onGloballyPositioned
+import androidx.compose.ui.platform.LocalDensity
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.unit.Density
+import androidx.compose.ui.unit.IntOffset
+import androidx.compose.ui.unit.IntSize
+import androidx.compose.ui.unit.dp
+import com.naver.maps.geometry.LatLng
+import com.naver.maps.map.compose.CameraPositionState
+import com.naver.maps.map.compose.CameraUpdateReason
+import com.naver.maps.map.compose.ExperimentalNaverMapApi
+import com.naver.maps.map.compose.LocationOverlay
+import com.naver.maps.map.compose.MapUiSettings
+import com.naver.maps.map.compose.Marker
+import com.naver.maps.map.compose.MarkerState
+import com.naver.maps.map.compose.NaverMap
+import com.naver.maps.map.compose.rememberFusedLocationSource
+import com.naver.maps.map.overlay.OverlayImage
+import com.teamoffroad.core.designsystem.component.StaticAnimationWrapper
+import com.teamoffroad.core.designsystem.theme.Black
+import com.teamoffroad.core.designsystem.theme.Sub2
+import com.teamoffroad.feature.explore.presentation.model.ExploreResultState
+import com.teamoffroad.feature.explore.presentation.model.LocationModel
+import com.teamoffroad.feature.explore.presentation.model.PlaceCategory
+import com.teamoffroad.feature.explore.presentation.model.PlaceModel
+import com.teamoffroad.offroad.feature.explore.R
+
+@OptIn(ExperimentalNaverMapApi::class)
+@Composable
+fun ExploreOffroadMap(
+    locationState: LocationModel,
+    places: List<PlaceModel>,
+    selectedPlace: PlaceModel?,
+    navigateToExploreCameraScreen: (Long, Double, Double) -> Unit,
+    updateLocation: (Double, Double) -> Unit,
+    updateTrackingToggle: (Boolean) -> Unit,
+    updateSelectedPlace: (PlaceModel?) -> Unit,
+    updateCategory: (String) -> Unit,
+    updatePlaces: (Double, Double) -> Unit,
+    updateCameraUiState: (ExploreResultState) -> Unit,
+    isValidDistance: (PlaceModel, LatLng) -> Boolean,
+    postExploreResult: (Long, Double, Double) -> Unit,
+) {
+    val density = LocalDensity.current
+    var markerOffset by remember { mutableStateOf(IntOffset.Zero) }
+    var mapViewSize by remember { mutableStateOf(IntSize.Zero) }
+    val backgroundPadding = 104
+
+    LaunchedEffect(locationState.cameraPositionState.cameraUpdateReason) {
+        if (locationState.cameraPositionState.cameraUpdateReason == CameraUpdateReason.GESTURE) {
+            updateTrackingToggle(true)
+        }
+    }
+
+    StaticAnimationWrapper {
+        Box(
+            Modifier
+                .fillMaxSize()
+                .padding(bottom = 74.dp)
+                .navigationBarsPadding()
+                .onGloballyPositioned { coordinates ->
+                    mapViewSize = coordinates.size
+                }) {
+            NaverMap(
+                properties = locationState.mapProperties,
+                uiSettings = MapUiSettings(
+                    isScaleBarEnabled = false,
+                    isZoomControlEnabled = false,
+                    isLogoClickEnabled = false,
+                    logoGravity = Gravity.TOP,
+                    // TODO: 릴리즈 이전에 로고 이미지 수정
+                    logoMargin = PaddingValues(top = 0.dp, start = 22.dp),
+                ),
+                locationSource = rememberFusedLocationSource(isCompassEnabled = true),
+                cameraPositionState = locationState.cameraPositionState,
+                onLocationChange = { location ->
+                    updateLocation(location.latitude, location.longitude)
+                },
+                onMapClick = { _, _ ->
+                    updateSelectedPlace(null)
+                },
+            ) {
+                LocationOverlay(
+                    position = locationState.location,
+                    icon = OverlayImage.fromResource(R.drawable.ic_explore_location_overlay),
+                    subIcon = locationState.subIcon,
+                    subIconWidth = 48,
+                    subIconHeight = 40,
+                    circleColor = Sub2.copy(alpha = locationState.circleAlpha),
+                )
+                places.forEach { place ->
+                    Marker(
+                        state = MarkerState(position = place.location),
+                        icon = OverlayImage.fromResource(R.drawable.ic_explore_place_marker),
+                        onClick = {
+                            markerOffset = calculateMarkerOffset(
+                                place.location, locationState.cameraPositionState, density, mapViewSize
+                            )
+                            updateSelectedPlace(place)
+                            updateCategory(place.placeCategory.name)
+                            true
+                        },
+                    )
+                }
+            }
+            ExploreMapForeground()
+            Box(
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .padding(top = (backgroundPadding + 18).dp)
+                    .align(Alignment.TopCenter),
+            ) {
+                if (locationState.isUserTrackingEnabled.not()) {
+                    ExploreRefreshButton(
+                        text = stringResource(R.string.explore_map_refresh),
+                        onClick = {
+                            updatePlaces(
+                                locationState.cameraPositionState.position.target.latitude,
+                                locationState.cameraPositionState.position.target.longitude,
+                            )
+                        },
+                        modifier = Modifier.align(Alignment.Center),
+                    )
+                }
+                ExploreTrackingButton(
+                    isTrackingEnabled = locationState.isUserTrackingEnabled,
+                    onClick = updateTrackingToggle,
+                    modifier = Modifier
+                        .align(Alignment.CenterEnd)
+                        .padding(end = 22.dp),
+                )
+            }
+            Row(
+                modifier = Modifier
+                    .align(Alignment.BottomCenter)
+                    .padding(bottom = 36.dp),
+                horizontalArrangement = Arrangement.Center,
+            ) {
+                ExploreMapBottomButton(
+                    painter = painterResource(R.drawable.ic_explore_quest_list),
+                    text = stringResource(R.string.explore_quests),
+                    onClick = {},
+                )
+                Spacer(modifier = Modifier.size(16.dp))
+                ExploreMapBottomButton(
+                    painter = painterResource(R.drawable.ic_explore_location),
+                    text = stringResource(R.string.explore_places),
+                    onClick = {},
+                )
+            }
+            ExploreAppBar(backgroundPadding)
+            selectedPlace?.let { place ->
+                Box(
+                    modifier = Modifier
+                        .fillMaxSize()
+                        .background(Black.copy(alpha = 0.25f))
+                        .pointerInput(Unit) {
+                            detectTapGestures(onTap = { updateSelectedPlace(null) })
+                        }
+                ) {
+                    Box(modifier = Modifier
+                        .align(Alignment.TopStart)
+                        .offset { markerOffset }) {
+                        ExploreInfoWindow(
+                            title = place.name,
+                            shortIntroduction = place.shortIntroduction,
+                            address = place.address,
+                            visitCount = place.visitCount,
+                            categoryImage = place.categoryImageUrl,
+                            onButtonClick = {
+                                when (isValidDistance(place, locationState.location)) {
+                                    true -> {
+                                        if (selectedPlace.placeCategory == PlaceCategory.CAFFE || selectedPlace.placeCategory == PlaceCategory.RESTAURANT) {
+                                            navigateToExploreCameraScreen(
+                                                place.id,
+                                                locationState.location.latitude,
+                                                locationState.location.longitude,
+                                            )
+                                        } else {
+                                            postExploreResult(
+                                                place.id,
+                                                locationState.location.latitude,
+                                                locationState.location.longitude,
+                                            )
+                                        }
+                                    }
+
+                                    false -> updateCameraUiState(ExploreResultState.LocationError)
+                                }
+                                updateSelectedPlace(null)
+                            },
+                            onCloseButtonClick = {
+                                updateSelectedPlace(null)
+                            },
+                            modifier = Modifier.align(Alignment.TopCenter),
+                        )
+                    }
+                }
+            }
+        }
+    }
+}
+
+private fun calculateMarkerOffset(
+    location: LatLng,
+    cameraPositionState: CameraPositionState,
+    density: Density,
+    screenSize: IntSize,
+): IntOffset {
+    val screenPosition = cameraPositionState.projection?.toScreenLocation(location)
+    return screenPosition?.let {
+        with(density) {
+            var xOffset = (it.x.toDp() - 115.dp).roundToPx()
+            var yOffset = (it.y.toDp() - 224.dp).roundToPx()
+
+            if (xOffset < 0) {
+                xOffset = 28
+            }
+            if (xOffset + 230.dp.roundToPx() > screenSize.width) {
+                xOffset = screenSize.width - 714
+            }
+
+            if (yOffset < 120.dp.roundToPx()) {
+                yOffset = 232
+            }
+            if (yOffset + 174.dp.roundToPx() > screenSize.height) {
+                yOffset = screenSize.height - 174
+            }
+
+            IntOffset(xOffset, yOffset)
+        }
+    } ?: IntOffset(0, 0)
+}
