+++ feature/explore/src/main/java/com/teamoffroad/feature/explore/presentation/model/ExploreUiState.kt
@@ -3,7 +3,7 @@ package com.teamoffroad.feature.explore.presentation.model
 data class ExploreUiState(
     val locationModel: LocationModel = LocationModel(),
     val places: List<PlaceModel> = emptyList(),
-    val authResultType: ExploreCameraUiState = ExploreCameraUiState.None,
+    val authResultType: ExploreResultState = ExploreResultState.None,
     val selectedPlace: PlaceModel? = null,
     val loading: Boolean = true,
     val isSomePermissionRejected: Boolean? = null,
@@ -13,13 +13,13 @@ data class ExploreUiState(
     val isUpdatePlacesFailed: Boolean = false,
 ) {
 
-    fun getExploreCameraUiState(errorString: String): ExploreCameraUiState {
+    fun getExploreCameraUiState(errorString: String): ExploreResultState {
         return when (errorString) {
-            ExploreCameraUiState.LocationError.toString() -> ExploreCameraUiState.LocationError
-            ExploreCameraUiState.CodeError.toString() -> ExploreCameraUiState.CodeError
-            ExploreCameraUiState.EtcError.toString() -> ExploreCameraUiState.EtcError
-            ExploreCameraUiState.Success.toString() -> ExploreCameraUiState.Success
-            else -> ExploreCameraUiState.None
+            ExploreResultState.LocationError.toString() -> ExploreResultState.LocationError
+            ExploreResultState.CodeError.toString() -> ExploreResultState.CodeError
+            ExploreResultState.EtcError.toString() -> ExploreResultState.EtcError
+            ExploreResultState.Success.toString() -> ExploreResultState.Success
+            else -> ExploreResultState.None
         }
     }
 }
